use the method position as the position of parameter sources
Summary: When marking a parameter of a method as a source and propagating the source, we attach the position of the first usage of the parameter to the taint. As discussed in T79810191, this can be confusing, and we would rather attach the position of the method signature instead. Once we have more accurate positions in the future, we could highlight the parameter name instead of the whole line.

Differential Revision: D25279831

fbshipit-source-id: f658a949fc0bb3f115d8b98fbdc93933d4aa6032