more accurate positions for methods
Summary: Right now, we use the position of the first instruction of a method as the position of the method itself. In 99% of cases, this is off by one since the method signature is on the previous line. Let's assume this is always true and substract one to the line position.

Differential Revision: D25279374

fbshipit-source-id: e16825521e7c5ae250bf47533066b53dd1fda5b2