inline methods returning an access path
Summary:
Following up on D25322586, this patch inlines methods that return a simple expression based on an argument.

This adds an attribute `inline_as` in models, which is either top, bottom or an access path. This is automatically inferred when analyzing a return statement, in the global fixpoint analysis. Note that it could also be supplied by a user using a model generator.

The inlining happens in the transfert function of the invoke instruction, if the callee model has the `inline_as` attribute and the model has no other side effects.

Differential Revision: D25344128

fbshipit-source-id: 20b4bdabba1a3252509ddc421a8738ff9fe37470