add tests to demonstrate false negatives due to aliasing
Summary:
This patch adds regression tests that demonstrate false negatives when a method returns an alias to a parameter, for instance a getter method.

This happens because we assume that each method call returns a new object in the heap.
```
Object getFoo() {
  return this.foo;
}
void issue() {
  getFoo().bar = source(); // get() returns a temporary object 1 which gets tainted
  sink(getFoo().bar); // get() returns a temporary object 2, untainted
}
```

In a follow-up, we will inline the call to getter methods, which should solve the issue.

Reviewed By: dkgi

Differential Revision: D25322586

fbshipit-source-id: f6b6c499a5f3af71fc30e19e11103e2447f5ca2a