implement a function that checks if a regular expression is a literal string
Summary:
This function checks if a regular expression is equivalent to a string equality, and if so, returns the string.
The plan is to use this later to optimize model generators, and specifically the `name` constraints that takes a regular expression. If the regular expression is a string literal, then we can potentially optimize the model generator by precomputing mappings (e.g, a map from a method signature to a method pointer).

Reviewed By: sinancepel

Differential Revision: D25352686

fbshipit-source-id: f633eef5fb8242981cb25774c89777320a29fca3