merge issues with the same rule
Summary: The current implementation merges issues with the same source kind, sink kind and rule. This leads to discrepancies in the Zoncolan UI, which relies on issues being merges when they have the same rule, regardless of source and sink kinds. This patch fixes the issue by using the same grouping strategy.

Differential Revision: D25502936

fbshipit-source-id: 7bda4ac02f8e587d47a7736e12b0e0c6aa522992