add a position to issues
Summary: This patch adds an explicit position to issues. Until now, the issue position was inferred in the zoncolan parser, using the maximum line number of the sinks. That means we would aggressively merge issues in callables, regardless of the line of the source or the sink. This patch uses the line of the sink as the issue line.

Reviewed By: dkgi

Differential Revision: D25508234

fbshipit-source-id: d0ed7c88f735646fca081d65fd296ad7457537f2