Fix highlighting issue where we highlight the wrong part of the line
Summary:
We currently only search for the first occurrence of the callee's name. This will highlight the wrong part of the line in situations like -
`String source = (String) Origin.source()`;
Where `source` is the callee.

Instead, we should search for callee + "(" so we are sure
we are highlighting a call of the callee.

(This may still give the wrong highlight if the same callee is called multiple times in a line, I have not seen instances of that yet..)

Reviewed By: arthaud

Differential Revision: D25625136

fbshipit-source-id: 9116aec1e36f80572c4f0df13d8b7f45b375ac56