only dump the base method in the call graph
Summary: This patch changes the `--dump-call-graph` to only dump the base callee instead of all possible callees (including overrides) of a method. That reduces the file size from 19GB to 1.2GB for fb4a. Users that do need the overrides can use `--dump-overrides` to get the override graph in a separate file.

Reviewed By: dkgi

Differential Revision: D25773324

fbshipit-source-id: c570002cb0a828639dd574d261858e3e9526d94d