Only augment positions in methods involved in issues
Summary:
Run a fixpoint to find all the frames involved in issues.

This will keep track of all the methods containing a frame that is in an issue and augment all the positions in that method's model (i.e. this can augment extra frames but won't augment extra models).

This improved the runtime of finding highlights on fb4a to around 30s

Reviewed By: arthaud

Differential Revision: D25702920

fbshipit-source-id: c83fca75ffede60a2541b613fb2239222be8d286