Rewrite child features to improve SNR
Summary: Currently a base class can be extended by multiple classes referenced in the AndroidManifest. This would mean that one exposed activity with permissions but another unexported activity would have the features all matched together. Therefore its better to simply and only attach a feature to the base class if there are no protections, permissions and the child is exported.

Reviewed By: arthaud

Differential Revision: D25883755

fbshipit-source-id: effffb42efab381d285ed6355065ddf7ef8e7966