Add support for Provider inline permission checks
Summary: If your Provider component extends specific base classes (like FbPermissionsContentProvider) or overrides a method like onCheckPermissions it means that the provider even when exported to third parties vi the manifest is still protected. This is because those permission checks are done in line. Therefore we check if these methods exist in the class or the class extends one of the classes.

Reviewed By: arthaud

Differential Revision: D25911288

fbshipit-source-id: dfa9d90659eae782e18244581053f0143d808b03