Return shallow list of unique constraints from a generator
Summary:
General High-Level Plan:
1. without looking at constraints, pre-calculate name->method set, class -> method set, class -> override method set
2. build int-> constraint lookup, then outer loop over methods (starting with fewer than all methods, using step 1 results wherever applicable) while caching constraint results in fast vector lookup
3. build a trie to further optimize regex matching; we can build the trie structure in preprocessing but use it as a lookup as an implementation detail of .satisfy regex methods.

This diff actually serves step 2, where we care about knowing all the constraints we're dealing with and want to un-nest them out of aggregate constraints like AllOf, AnyOf, and Not.

Ended up just finishing this implementation to try to learn C++ rather than scrap and start with the precalculated method sets, but will be moving onto that immediately after this.

Reviewed By: arthaud

Differential Revision: D26376001

fbshipit-source-id: 51cbdadd61d4cef696d6abac406d48fcda71148f