Address false positive Fragment::onActivityCreated
Summary: When triaging I realized Fragment::onActivityCreated https://android.googlesource.com/platform/frameworks/support/+/androidx-1.0-dev/fragment/src/main/java/androidx/fragment/app/Fragment.java#2463 which can potentially delegate to any fragment that overrides onActivityCreated. This should not be the case because the route would be very hypothetical. In the future we should consider onActivityCreated as a source when we have more information on when a fragment is exported.

Reviewed By: yuhshin-oss

Differential Revision: D26484820

fbshipit-source-id: ef0c2c88e1b3566039136c1fb4d8e0d193f8bca5