Create feature where user input flows via query parameter
Summary: In cases where a URL is user controlled an used in a Webview we usually don't care much if the entire URL is reused since it would already require you to load that URL and domain however a query parameter can be set to anything and is therefore more dangerious. This feature allows us to include only those flows.

Reviewed By: yuhshin-oss

Differential Revision: D26485499

fbshipit-source-id: 17b01de1bf4d047c991dac109e54f461a40493a2