Exclude common Service AIDL false positives
Summary:
When I was looking through different sources I realized that our AIDL sources are not restricted enough. For example constructors are not sources, we do not care about google, android or chromium AIDL services and two specific examples generated a lot of false positives. So filtering these all out.

Finally when we map a AIDL stub back to the actual service we should not only check for Service.*Handler but also for Service.*Stub which I added.

Reviewed By: yuhshin-oss

Differential Revision: D26484637

fbshipit-source-id: caa861e32a38b40eeaf239045894889232a321c2