Use namespace instead of class for mapping generators
Summary:
rfc: is this preferable to having a class full of static methods?

Not sure about what good cpp style is.. alternatively I can change the constructor to stop taking context, or change the static methods to stop requiring methods to be passed in and just read from the class state.

Reviewed By: arthaud

Differential Revision: D26519327

fbshipit-source-id: 15cada64d03618d47d79b4329fa7b1352258bb94