Add via_type_of field to `Frame`
Summary:
Implementing via-type-of feature similar to pysa, in order to catch Independent flows: https://fb.quip.com/DHA5Aa6Hebwb
Via-type-of features will be user-specified as a set of ports we want to propagate as "via-type-of:(type seen at callsite during analysis)" features. They will be stored in source or sink frames as specified in model generators and attached to any taint flowing through the callsite in that frame.

This diff adds a field to store via_type_of in `Frame` and updates all the calls to its constructor

Reviewed By: arthaud

Differential Revision: D26378694

fbshipit-source-id: 71ba1b7ff3212e4e106d782cdb190ea65401bd5d