Adding optimistic inference to ClassProperties whether a Fragment is user controlled
Summary:
Please also review the larger idea of this detection pattern.

The basic version is that we know based on whether a class is exported whether its user controlled. The problem is that because we can't do multi component analysis (yet) we don't know if any Fragment is user controlled since the only way those are called is via activities. For example the following pattern
```
ClassAExported {
  function oncreate() {
    Intent intent = getIntent() // source
    Fragment fragment = FragmentInstance.Class
    fragment.setArguments(intent)
  }
}

FragmentInstance {
  getArguments() // second source, which would have the exported feature attached after this diff
}
```

What this diff does is use a Regex to parse out all fragments referenced in the bytecode of an exported Activity and add to them the feature via-caller-exported. This allow us when setting up filters to know which Fragment instances using getArguments() to prioritize since if the activity is exported and user controlled the fragment will likely be (semi-)user controlled as well.

Reviewed By: yuhshin-oss

Differential Revision: D26518272

fbshipit-source-id: 2beb41c05d99cf9d5756fdff9aaff6408e9891b3