Introduce OkHttp Cert Pinning rule
Summary:
Adding a new rule for OkHttp clients not having certificate pinning. I added cert pinning as a feature we can filter in the rule.

There is some nuance to triage the issue as well where we only care about cert pinning comms with FB services but that should be clear to the oncall anyway.

Reviewed By: GerbenJavado

Differential Revision: D26182120

fbshipit-source-id: 5259f6be7758e0308db371a734bdd13fcef09371