Rewrite WebViewModelGenerator and remove duplicate
Summary: My initial plan was to remove Webview::evaluateJavascript as its already defined in the ExecuteJavascript generator and then I came across this very verbose generator so decided tor rewrite it as well. The reason we dont have strict signatures here but the parent and super lookup is because a lot of webviews extend each other and this way we can have the sink in the webview instance and skip-analysis on all the base classes to prevent the analysis from going in circles. There are still further improvements to be made here but one step at the time.

Reviewed By: yuhshin-oss

Differential Revision: D26578034

fbshipit-source-id: d6b6b75b1e14d9bd71f06d2d36cb86e404e6a390