Create dedicated rule for server response to HTML parser
Summary: Server responses flowing into HTML parser was part of the larger catageory 4028. This is useful for debugging exactly what flows we have from the server but makes it hard to productionize an individual sink kind due to grouping, documentation being generic etc. So since I think we can productionize this, separating it out. For performance we could also consider opting 4028 out of diff analysis.

Reviewed By: yuhshin-oss

Differential Revision: D26578469

fbshipit-source-id: 0dde011aa66e9f6b54e42967e0caa51ef93ec67f