Add method filtering structure on constraints
Summary:
Structure method filtering as a method on each constraint. The constraint will return either Top, or a filtered set if we override the implementation (ie., we will do this for method name constraint, etc.).

We'll also override this for aggregate constraints like any, all, or not to call `may_satisfy` on their children and do intersections/unions or some other heuristic to start.

The JsonModelGenerator can simply run `may_satisfy` once on its all-encompassing all-constraint to get a method subset to pass into `run`.

Reviewed By: yuhshin-oss

Differential Revision: D26537465

fbshipit-source-id: 1ec047abf93ff15b9bd590356157934abae89060