Implement may_satisfy filtering for type constraints
Summary:
This was a bit of a struggle to implement cleanly due to the object oriented nature of constraints and cpp lack of good support for isinstance equivalent.

Essentially to make use of the class-to-methods mapping and class-to-override-methods mapping, we want to narrow the may_satisfy method pool when we have a combination of constraints.

ie., for class-to-methods, we want to do a lookup if we have a ParentConstraint wrapped around a TypeNameConstraint.  for class-to-override-methods, we want to do the lookup if we have a ParentConstraint wrapped around an ExtendsConstraint wrapped around a TypeNameConstraint.

I made some decisions under the assumption that TypeNameConstraints are the only type of may_satisfy optimization that we plan to implement. I've added `may_satisfy` to the type constraint interface and will return top in all cases except the type name constraint case. We assume that the only caller of this function will be the ParentConstraint MethodConstraint class.

Furthermore, after a lot of deliberation, I'm using a boolean flag to determine whether or not we passed through an Extends wrapper and should therefore use the class-to-override-methods lookup. Suggestions are super welcome on how to do this in a cleaner way in cpp.

Reviewed By: yuhshin-oss

Differential Revision: D26583483

fbshipit-source-id: 057981c56a763737d5328c3d4cd3c82d4c5755aa