Materializing and propagating via-type-of features for argument ports
Summary: `FrameSet::propagate` now materializes any via-type-of ports it sees and includes them in the inferred features. It does this by taking in an additional argument containing the `DexType`s of its (argument) ports that were seen at runtime. This will allow the analysis at callsites to use models with the propagated via-type-of features.

Reviewed By: yuhshin-oss

Differential Revision: D26502031

fbshipit-source-id: 030d7a5265aeca4f150a04aa213600596b55c326