ignore parameter type overrides in the signature constraint
Summary:
The current implementation of the signature constraint uses `method->show()`, which adds the parameter type overrides.
This diff adds a getter for the signature of a method, which is different than the `show()`.
For instance, `show(m)` could be `Ljava/Foo;.bar(Ljava/Foo;)V[0:Ljava/Bar;]` but `signature(m)` is `Ljava/Foo;.bar(Ljava/Foo;)V`

Reviewed By: shannonzhu

Differential Revision: D26732225

fbshipit-source-id: e78765691d3183cc808bca9350e5adb5027004e6