Parallelize creation of method mappings
Summary:
Trying to profile what is taking so long with creating method mappings, and it looks like the mapping creation is taking a surprisingly tiny number of cycles (some 0.3% of total) even when profiling the work done specifically targeted at this window: https://fburl.com/strobelight/jimuknpi

This leads me to believe that the reason this takes proportionally quite long in the wall time is because the work done is super easy but everything is essentially sequential - we're grabbing a lock on the method mapping for each method so wins from parallelizing within each mapping creation function is minimal (in fb4a, with/without concurrency is a delta of about 5% mapping generation runtime).

Instead, I tried parallelizing the creation of the 3 maps themselves.

Was hoping for a bigger delta here, but this gives a 6% speedup in generating mappings regardless of app size. Not nothing, so might as well put this up for now.

Reviewed By: arthaud

Differential Revision: D26782128

fbshipit-source-id: 2180cb33d9784cf8728d80c78db00d052e6c3dad