Add feature for flows via this.this$0
Summary:
Getting comments about the approach which is an attempt to address the FP (let's worry about the impl later).

FP: https://www.internalfb.com/intern/zoncolan/122696863/?db_override=xdb.mariana-trench-fbsource&issue_instance_id=122973003

The noise is coming from the trace frame that says "com.facebook.video.plugins.MatureContentRatePillPlugin::setToolTip argument(0)":

The first argument of `SecureHtml.fromHtml` does lead to the sink. However, in this case, argument(0)/this of setToolTip does not provide any meaningful taint flow into the sink:

https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbandroid/java/com/facebook/video/plugins/MatureContentRatePillPlugin.java?commit=109797f1fd9cd32bac0d6e566c280bd12f5d3341&lines=141%2C149-156

"This" flows into the constructor of the anonymous class and gets set on a field with name `this$0` which looks like (see line 7):
{P222506784}

Capturing the outer-class "this" object seems logical because one could easily use "outerClass.mField" within the anonymous class, but as in the issue reported, it has potential noise.

**TLDR; This diff adds a breadcrumb (via-anonymous-this) for flows going via "this$0" fields which can be filtered out (see test plan).**

Reviewed By: arthaud

Differential Revision: D26707801

fbshipit-source-id: 7fbbaa90b6bad559d2efb44603088e8d737699ef