Multi-source: Track counterpart sources in TriggeredPartialKind
Summary:
Previously, a multi-source issue is created upon detection of a flow: `lblA:SourceA -> TriggeredPartialSink(lblA)`. This does not take into account the rule that materialized `PartialSink(lblA)` to `TriggeredPartialSink(lblA)` and causes FPs when partial sink names are re-used.

Example, consider the following rules that differ only by the sources in "lblA":
```
    //   Code: 1000
    //   Sources: { lblA: [SourceA], lblB: [SourceB] }
    //   Sinks: [ Partial(SinkX, lblA), Partial(SinkX, lblB) ]
    //
    //   Code: 2000
    //   Sources: { lblA: [SourceC], lblB: [SourceB] }
    //   Sinks: [ Partial(SinkX, lblA), Partial(SinkX, lblB) ]
    //
    // This flow would satisfy both rules, even if Triggered(SinkX, lblB) was
    // only created from seeing `SourceA -> Partial(SinkX, lblA)`:
    //
    //   SourceB -> Triggered(SinkX, lblB)
```

To fix this, each TriggeredPartialKind needs to track the rule that caused it to be triggered. That way, the flow in the above example would look like the following and it becomes clear which rule was satisfied.

```
    //   SourceB -> Triggered(SinkX, lblB, rule: 1000)
```

Reviewed By: arthaud

Differential Revision: D27284046

fbshipit-source-id: f0ca985dd28825af0b0c7ebac575b6954e4b3d24