Multi-source: Track fulfilled partial kind state in its own class
Summary:
For multi-source flows, whenever Transfer.cpp detects that one of the two flows are fulfilled, it needs to track state related to that flow. Currently, it only tracks "features" of that flow.

In order to address a bug that occurs when a partial sink name gets reused (leads to FPs), we should also track which flow in the multi-source rule was satisfied.

More details to come with the actual fix itself. Meanwhile, this introduces a FulfilledPartialKindState class which acts as a container for tracking state related to fulfilled partial sinks. It currently only has "features". More state to be added.

Reviewed By: SimranVirk

Differential Revision: D27284047

fbshipit-source-id: 49d711a891626e9634ccbfb0fb37ef7d9ee28e48