Avoid returning void on non-void test functions
Summary:
Turning on global analysis was crashing many of our unit tests with an instruction source indexing exception.

Additional context on where the redex exception is thrown and why: D27799390

MT tests create dummy functions with a non-void return type, but for test simplicity the body just contains a void return. Redex global analysis crashes because it assumes the non-void return type implies it's always safe to index into the first argument of the return instruction.

To fix this, let's return a base `Object()` type in the function body of non-void test functions. Open to other suggestions for any cleaner options.

Reviewed By: arthaud

Differential Revision: D27801619

fbshipit-source-id: d30b13eab007867d1c71c09ca29808f002b3c8ff