add a script to run mariana trench without running the setup.py
Summary: This patch adds a script `cmake_shim.py` that is compiled into `build/mariana-trench`. This allows to run MT with local changes (Python or C++) without having to re-run the `setup.py`. It takes care of setting the PATH and PYTHONPATH correctly. That should be helpful for open source users trying to contribute.

Reviewed By: dkgi

Differential Revision: D27901201

fbshipit-source-id: a05ac48535e6ed7a84aaf322112658afad855f87