Support parsing interfaces as non methods in integration tests
Summary:
Handle parsing integration tests with abstract methods. These methods were previously being added to redex with no distinction from non-abstract methods. The problem is, abstract methods may have a non-void return type, but the body must return void.

This causes crashes when we turn on redex global type inference, which includes source code that assumes it can check the return annotation of a non-abstract function and then index into the return instruction if it is not void.

I don't think there is any viable workaround or even hack that we can use to fill the dalvik bytecode body of the abstract method. The only option is to parse out which class methods are part of interfaces and build the redex class accordingly.

Ran into some issues due to the fact that redex classes can only be created once, and we have to know what methods are present when we create. If we want to create methods and attach this to an interface class, we need to do so before we create that interface class with no methods (because we need to instantiate the super type of another class we saw). Additionally, the MT redex library cannot currently support creating a class that has both abstract and concrete methods.

Reviewed By: yuhshin-oss

Differential Revision: D27874263

fbshipit-source-id: 1fb64f9abd4b48d707a7eb7c398dcce2c6427182