Remove method bodies of abstract classes without setting external
Summary: More cleanly set a dex method we're creating as abstract without having to make it non-concrete and external. It turns out the opposite of concrete is not abstract; concrete or not is referring to whether the method actually exists as opposed to is just something being referenced. See: https://fb.workplace.com/groups/redex.feedback/permalink/3740919949311095/

Reviewed By: yuhshin-oss

Differential Revision: D27901043

fbshipit-source-id: bb3d44991740f40ade2896e574c47b95cb4afe02