Apply annotation and abstract settings per method on create
Summary:
Overloading the create_methods functionality here so we can avoid changing most call sites that just want to provide a set of method body strings and don't want to worry about setting any or all of them as abstract or with annotations.

The remaining callsites can invoke the new create_methods, which actually associates the annotations (and the new abstract flag) with a specific method rather than applying every annotation(s)/abstract flag to each method in the class.

Note: because we cannot add to methods after we call .create() on the dex class creator, we can't build this in a way that adds one method at a time. This will also clean up D27874263, which I will rebase on top.

Reviewed By: yuhshin-oss

Differential Revision: D27901047

fbshipit-source-id: 54b146dad250c9b8463593ea4c9ab27fc3575ce2