Use hashed set abstract domain for method mapping sets
Summary:
Copying the diff summary from D28019267, which this diff is replacing. The only difference is that we are using HashedSetAbstractDomain instead of changing MethodSet to use an unordered_set instead of a patricia tree. MethodSet is still used as a data structure for storing leaves in the taint representation, where speed for joins is more important.

Note: I've also stacked this on top of D28020550, which re-introduces override mappings and further reduces runtime.

----

I think I've finally gotten to the bottom of why method mappings have delivered such incredibly lukewarm performance changes, despite bottoming out the .satisfy runtime of over 1/3 of all model generators to 0.

**tl;dr** - Patricia Sets were nice for very fast joins/meets/differences between filtered method sets, but were tanking the performance of the .satisfy run due to a very complicated iteration increment function.

Longer version:

Signature satisfy is still the longest running constraint type, and I was looking into adding a mapping that stores signatures by prefix to narrow the search space for this constraint in particular.

When doing some more granular perf logging, I noticed that the same .satisfy function was consistently taking *longer* to run on certain model generators, even though the search space had successfully been reduced. There was no functional difference between the two .satsify methods called - the only difference was the method set:

Without the new mapping: P410792027
With the new mapping: P410799775

If you notice in particular,

```
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run on `BloksSinkGenerator` in: 0.52s ... checked 964647 methods.
===DONE==== BloksSinkGenerator: -1
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run on `BloksSinkGenerator` in: 0.52s ... checked 964647 methods.
===DONE==== BloksSinkGenerator: -1
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run on `BloksSinkGenerator` in: 0.54s ... checked 964647 methods.
===DONE==== BloksSinkGenerator: -1
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run on `BloksSinkGenerator` in: 0.61s ... checked 964647 methods.
```

We're checking 960k methods here, and finishing significantly faster than when we check 800k methods instead:

```
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run_filtered on `BloksSinkGenerator` in: 0.92s ... checked 799716 methods.
===DONE==== BloksSinkGenerator: 799716
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run_filtered on `BloksSinkGenerator` in: 0.92s ... checked 799716 methods.
===DONE==== BloksSinkGenerator: 799716
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run_filtered on `BloksSinkGenerator` in: 0.96s ... checked 799716 methods.
===DONE==== BloksSinkGenerator: 799716
Checked may_satisfy on `BloksSinkGenerator` in: 0.00s
Finished run_filtered on `BloksSinkGenerator` in: 1.06s ... checked 799716 methods.
```

These measurements didn't include any of the method set size calculations, only a single-line call to `.run`, which grabs the `.begin` and `.end` from the container passed in, and shells out to logic that is identical regardless of whether we're using filtered methods or not (logging D27629414).

This made me realize that the fact that unfiltered methods was a `Methods` container and the filtered methods was a `MethodSet` container was probably the root cause of the difference. Patricia Sets were nice for very fast joins/meets/differences between filtered method sets, but were tanking the performance of the .satisfy run due to a very complicated iteration increment function.

This diff changes the underlying implementation to use a standard unordered set. We are making a trade-off, accepting slower set joins/meets/difference calculations and gaining faster iteration. This is totally worth it because joins/meets/etc. only happen on AllOf, AnyOf and Not constraints, and often times are dealing with already minimized sets of functions. The full iteration happens for every method generator and over a lot of methods.

Reviewed By: arthaud

Differential Revision: D28072237

fbshipit-source-id: f66a715eafd195ac3904974050bd12cd3ace78bd