CRTEX: Callee_port must be provided for "Producer" ports, not inferred
Summary:
Fixed a bug in CRTEX frame validation `validate_and_infer_crtex_callee_port`.

Previously, the validation logic infers the port as either "Anchor" or "Producer" based on the canonical names provided.

This fix changes it to only infer "Anchor" if a callee port wasn't given (and canonical names are templated). It should use the user-provided callee_port otherwise.

This was causing an issue when CRTEX consumers provide a callee port of the form "Producer.<producer id>.<canonical_port>" and the validation logic overwrote that to just "Producer", which broke traces that relied on <canonical_port>.

Reviewed By: dkgi

Differential Revision: D28425699

fbshipit-source-id: 1311e8dec01d89b739e9dbaa828bf307f297e883