Fail if multiple life-cycle definitions share the same method name.
Summary: Life-cycle definitions consist of a base class, a method name, and a list of callees. For every final child of base class, a method of the form `ChildClass.<method_name>` is created. If `ChildClass` extends multiple base classes with life-cycle definitions with the same method name, this method is created twice which causes a conflict. To avoid getting into this situation, method names are required to be unique across all life-cycle definitions.

Reviewed By: arthaud

Differential Revision: D28655647

fbshipit-source-id: 9e1091cdcd93189c5e5858fe36387c9bb4f9792e