Use default values for Frame() constructor in tests
Summary:
Whenever we add a new field to `Frame`, it gets initialized from an argument in the constructor. We then have to update a whole lot of tests that call the constructor.

We want to avoid using default values because we should revisit all `Frame()` call-sites and make sure the newly added field is initialized correctly. However, this does not apply to tests.

Here's we add a `test::make_frame(kind, FrameProperties)`. `FrameProperties` is a struct that has all Frame fields initialized to a default value. Individual test cases will override specific fields only when needed. When a new field is added to `Frame` in the future, we only need to update this method and specify a suitable default.

Reviewed By: dkgi

Differential Revision: D28697332

fbshipit-source-id: 7a4aae3aa0b52ece5d240d7fef04d9ca9f3026bc