Widen strip_subclass to apply more manifest features
Summary: There are still various components that are present in the manifest but not recognized as such by Mariana Trench. This is because a reference in the manifest to `MessagingService` can result in a source like `MessagingService$binder$1`. The previous logic would try and handle this case but was too strict. Approximately 149 issues did not have features https://fburl.com/zoncolan/vuzx996p. I refactored this logic to be more genrious in it matching but also match against the exact class if the subclass is mentioned in the manifest. This is still not perfect as we can have in the manifest `Service$Receicer;` And the source `Service$Receiver$1;` which would only match back to Service but Im confident this logic is an improvement over the current use.

Reviewed By: dkgi

Differential Revision: D29228155

fbshipit-source-id: 49d304d9c5bfadebcaf8dca496064f0b3b45b8e2