Overwrite local with global type instead of joining
Summary:
There are likely a couple reasons why running redex global type analysis is not resulting in meaningful changes in MT issues output.

Firstly, we need to start passing in proguards and making sure the global type analysis is not treating every function as a possible entry point. This would prevent any refinement on parameter types even if all callsites pass in a narrower type, which is specifically the global analysis refinement we're hoping to see. Will be looking into proguards as a follow-up to this diff.

Even with entry points specified, the logic to apply global type analysis should not be joining local and global types - we'd lose all refinement information we get here. Sometimes, redex global type analysis cannot do better than local analysis and will give up, returning top. To handle this edge case, we can just check for top and fall back to local if we see it, otherwise simply overwrite the local type with the globally inferred one.

Reviewed By: arthaud

Differential Revision: D30504887

fbshipit-source-id: 601b45f9c90c7ded5b37d359e129b312499e3f35