Gate unreachable code pruning behind shim flag
Summary:
D19838403 introduced the possibility of passing in a proguards configuration file, and upon doing so, removing all unreachable code from the analysis.

This flag is not being used anywhere, so I'm assuming the default behavior we want is to take the extra execution time and still report potential vulnerabilities in code that is unreachable but checked in. Perhaps RFC on what the default should be here.

Either way, without proguards specifying entry points, we're not able to run meaningful global type analysis with redex. I'd like to be able to gate the removal of unreachable code behavior to avoid contaminating runs with and without proguards passed in. I imagine being able to run with and without this pruning is meaningful in the future.

Reviewed By: dkgi

Differential Revision: D30531686

fbshipit-source-id: 4e0872a4d5af50500fddfa97f96492b8471c2219