Don't taint-in-taint-this on Component Classes
Summary:
A common false positive is when Mariana Trench taints Argument(0) on an obscure method if any of the passed in arguments is tainted. For example

`super.shouldOverrideUrlLoading(tainted)` would also taint the super/this position. This is incorrect. Therefore adding a denylist that methods should not propagate taint to this when this is a main component.

Reviewed By: SimranVirk

Differential Revision: D32062883

fbshipit-source-id: 102af86f30d45d25bc14d2c98885d1219105dba8