Don't attach via-obscure-taint-in-taint-this on setters
Summary:
`via-obscure-taint-in-taint-this` has the potential of being a good feature to detect false positives where we overtaint Argument(0) due to missing code. The problem is that we also attach this feature on flows like
```
intent.putExtra(tainted)
```
Where Argument(0) should be tainted. This is a first approach at not applying `via-obscure-taint-in-taint-this` when we deal with a setter/writer where normally Argument(0) should get tainted.

To make sure that every Model() constructor avoids adding via-obscure-taint-in-taint-this moving this logic to the model generator.

Reviewed By: SimranVirk

Differential Revision: D32020635

fbshipit-source-id: 1f0cc34952566bdcb72a0e4bf6d5fed7d6134d8b