Support via-value-of
Summary: This diff adds support for `via_value_of` features in the analysis. This tracks the constant value arguments (int literals and strings)  at callsites and materializes them as `via-value:<Constant value at callsite>` features in the associated model. When the arguments are not constant values, the feature will show up as `via-value:unknown`.

Reviewed By: yuhshin-oss

Differential Revision: D32656474

fbshipit-source-id: 30a5d99deb4a91cfc1fc7de066430dd3e4019bf6