Map private classes to source file
Summary: In Kotlin when a (private) class is declared the underlying logic will actually treat this as an independent class which results in Mariana Trench seeing the callable as package.name.PrivateClass and cause it to fail mapping it to the right source file. This can be addressed by also mapping private classes.

Reviewed By: dkgi

Differential Revision: D33163475

fbshipit-source-id: 37cfe4f563041fbab2f2863de116bdf50b5d76fa