Update expected output of integration tests
Summary:
hg bisect showed that this change in expected output was caused by D34052846 which made changes to the MethodCreator which is used by lifecycle methods - https://www.internalfb.com/code/fbsource/[526348ea5409]/fbandroid/native/mariana-trench/source/LifecycleMethod.cpp?lines=156. Not fully sure how this results in us getting a position with line 0, but it seems like a harmless difference.

(In the new expected positions, the positions are added to frames that are included in issues (this is as expected based on our highlights logic) and they show up without a file and in line 0 because they are within the activity_lifecycle_wrapper function. )

Reviewed By: yuhshin-oss

Differential Revision: D34149760

fbshipit-source-id: 2aaf789ca3ef83a8eb1cfe047b4d05513dba573c