CallPositionFrames: Propagate
Summary:
The `propagate` operation needs to operate on a per-kind basis. Currently, this happens in `FrameSet`. In the new model, this should happen in `CallPositionFrames` where the kind is known.

Heavily based on `FrameSet` with one minor difference. In `CallPositionFrames` we handle multiple kinds, in `FrameSet`, there is only one kind.

Reviewed By: SimranVirk

Differential Revision: D34232709

fbshipit-source-id: a10875fc4808a2ad821be47dea29985f21d5776f