CallPositionFrames: show() support
Summary:
Part of the changes for moving Taint structure from `Kind -> Callee -> Position -> Frames` to `Callee -> Position -> Kind -> Frames`

Overriding `operator<<` for better debugging output. Unlike most other `CallPositionFrames` methods which closely mimic `FrameSet`, this one is different because it is intended to reflect the internal structure of the object for debugging purposes.

Reviewed By: SimranVirk

Differential Revision: D34426756

fbshipit-source-id: 2f3a249e79a01575a9a487103f1850b867f89c42