Using dependency graph for class properties
Summary: Mariana Trench adds features using information from Manifest file such as if the class is exported (i.e. User Controlled) in issues. We are missing these features when issues are found in classes which are not directly configured in the manifest file  (eg. util classes). These can be captured by checking if one of the known exported classes is reachable from the utility methods using the call graph. This diff adds a simple bfs traversal of the dependency graph to add these missing features. `ClassProperties` now stores a reference to the dependency graph for this.

Reviewed By: yuhshin-oss

Differential Revision: D34053066

fbshipit-source-id: eb33f65f1345a13dfbbb883ba24397b29573da21